+++ CHANGELOG.md
@@ -12,7 +12,8 @@ changes.
 
 ### Added
 
--
+- Unsancho GovTool [Issue 1637](https://github.com/IntersectMBO/govtool/issues/1637)
+- Add network metrics model to frontend service
 
 ### Fixed
 
@@ -27,7 +28,6 @@ changes.
 ### Added
 
 - Add network name to GET /network/metrics [Issue 1644](https://github.com/IntersectMBO/govtool/issues/1644)
-- Add network metrics model to frontend service
 
 ### Fixed